requirements.txt
# TextCNN 文本分类项目依赖

# 深度学习框架
torch>=2.0.0

# 数据处理
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# 中文分词
jieba>=0.42.1
synonyms>=0.0.4  # 同义词替换

# 配置管理
pydantic>=2.0.0
pyyaml>=6.0

# Web API
fastapi>=0.100.0
uvicorn>=0.23.0

# 日志和可视化
tensorboard>=2.13.0
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0

# 模型导出
onnx>=1.14.0
onnxruntime>=1.15.0

# 工具库
loguru>=0.7.0
